+++ tiger-testenv-mgr/src/frontend/src/types/Features.ts
@@ -1,12 +1,12 @@
 /*
  * ${GEMATIK_COPYRIGHT_STATEMENT}
  */
-export class ExperimentalFeatures {
+export class Features {
   trafficVisualization: boolean = false;
 
 
-  static fromMap(map: { [x: string]: string; }): ExperimentalFeatures {
-    const features = new ExperimentalFeatures();
+  static fromMap(map: { [x: string]: string; }): Features {
+    const features = new Features();
     features.trafficVisualization = map["trafficvisualization"] === "true";
     return features;
   }